pyproject.toml
[project]
name = "movie-rag-system"
version = "0.1.0"
description = "A RAG system for movie reviews and comments from multiple sources"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.28.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.0.29",
    "chromadb>=0.4.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "selenium>=4.15.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "streamlit>=1.28.0",
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
    "httpx>=0.25.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"
